# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HACCP Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tenant {
  id: number;
  name: string;
  createdAt: string;
}

export interface Market {
  id: number;
  tenantId: number;
  name: string;
  code: string;
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  label: string;
  sortOrder: number;
  createdAt: string;
}

export interface Section {
  id: number;
  categoryId: number;
  number: string;
  title: string;
  description?: string | null;
  sortOrder: number;
  createdAt: string;
}

export type FormDefinitionFieldType =
  (typeof FormDefinitionFieldType)[keyof typeof FormDefinitionFieldType];

export const FormDefinitionFieldType = {
  temperature: "temperature",
  boolean: "boolean",
  text: "text",
  photo: "photo",
  signature: "signature",
  pin: "pin",
} as const;

export interface FormDefinition {
  id: number;
  sectionId: number;
  fieldName: string;
  fieldType: FormDefinitionFieldType;
  label: string;
  required: boolean;
  validationMin?: number | null;
  validationMax?: number | null;
  warningThreshold?: number | null;
  sortOrder: number;
  createdAt: string;
}

export type FormInstanceStatus =
  (typeof FormInstanceStatus)[keyof typeof FormInstanceStatus];

export const FormInstanceStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface FormInstance {
  id: number;
  marketId: number;
  sectionId: number;
  year: number;
  month: number;
  status: FormInstanceStatus;
  createdAt: string;
}

export interface CreateFormInstance {
  marketId: number;
  sectionId: number;
  year: number;
  month: number;
}

export interface FormEntry {
  id: number;
  formInstanceId: number;
  formDefinitionId: number;
  value?: string | null;
  signature?: string | null;
  pin?: string | null;
  photoUrl?: string | null;
  entryDate: string;
  createdAt: string;
}

export interface CreateFormEntry {
  formDefinitionId: number;
  value?: string | null;
  signature?: string | null;
  pin?: string | null;
  photoUrl?: string | null;
  entryDate: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  SUPERADMIN: "SUPERADMIN",
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export interface User {
  id: number;
  tenantId: number;
  name: string;
  email: string;
  role: UserRole;
  initials?: string | null;
  pin?: string | null;
  createdAt: string;
}

export type ListMarketsParams = {
  tenantId?: number;
};

export type ListFormInstancesParams = {
  marketId: number;
  sectionId?: number;
  year?: number;
  month?: number;
};

export type ListUsersParams = {
  tenantId?: number;
};

export type SeedData200 = {
  message: string;
};