# lib/api-zod/src/generated/types/formInstance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HACCP Management API
 * OpenAPI spec version: 0.1.0
 */
import type { FormInstanceStatus } from "./formInstanceStatus";

export interface FormInstance {
  id: number;
  marketId: number;
  sectionId: number;
  year: number;
  month: number;
  status: FormInstanceStatus;
  createdAt: Date;
}