# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HACCP Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Market {
  id: number;
  tenantId: number;
  name: string;
  code: string;
  createdAt: Date;
}