# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HACCP Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  name: string;
  createdAt: Date;
}