# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HACCP Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  tenantId: number;
  name: string;
  email: string;
  role: UserRole;
  initials?: string | null;
  pin?: string | null;
  createdAt: Date;
}